requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
Pillow==10.0.1
google-generativeai==0.3.1
python-dotenv==1.0.0
scikit-learn==1.4.0
pandas==2.2.1
numpy==1.26.4
joblib==1.3.2 
